% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_core.R
\name{add_mcview_config}
\alias{add_mcview_config}
\title{Add MCView configuration to DAF object}
\usage{
add_mcview_config(
  daf_obj,
  title = NULL,
  tabs = NULL,
  light_version = FALSE,
  excluded_tabs = NULL,
  cache_in_daf = NULL,
  cache_daf_root = NULL
)
}
\arguments{
\item{daf_obj}{DAF object}

\item{title}{Application title}

\item{tabs}{Character vector of tabs or comma-separated string}

\item{light_version}{Whether to use light version}

\item{excluded_tabs}{Character vector of excluded tabs}

\item{cache_in_daf}{Whether to store cache in the DAF}

\item{cache_daf_root}{Root path for cache DAF storage}
}
\description{
Add MCView configuration to DAF object
}
