% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_accessors.R
\name{add_optional_vec}
\alias{add_optional_vec}
\title{Add an optional vector to a tibble}
\usage{
add_optional_vec(tbl, daf_obj, axis, vec_name, col_name = vec_name)
}
\arguments{
\item{tbl}{Tibble to add vector to}

\item{daf_obj}{DAF object}

\item{axis}{Axis name}

\item{vec_name}{Vector name}

\item{col_name}{Column name in result (defaults to vec_name)}
}
\value{
Modified tibble with vector added if it exists
}
\description{
Helper to reduce repeated pattern of checking and adding optional vectors.
}
