% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_accessors.R
\name{add_optional_vecs}
\alias{add_optional_vecs}
\title{Add multiple optional vectors to a tibble}
\usage{
add_optional_vecs(tbl, daf_obj, axis, vec_names)
}
\arguments{
\item{tbl}{Tibble to add vectors to}

\item{daf_obj}{DAF object}

\item{axis}{Axis name}

\item{vec_names}{Character vector of vector names}
}
\value{
Modified tibble with existing vectors added
}
\description{
Helper to add multiple optional vectors in one call.
}
