% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gene_mc.R
\name{calc_module_correlations}
\alias{calc_module_correlations}
\title{Calculate correlations for genes as a module/anchor}
\usage{
calc_module_correlations(
  dataset,
  genes,
  n_top = 30,
  cell_type_filter = NULL,
  atlas = FALSE
)
}
\arguments{
\item{dataset}{Dataset name}

\item{genes}{Vector of gene names to treat as module}

\item{n_top}{Number of top correlations to return}

\item{cell_type_filter}{Optional vector of cell types to filter metacells}
}
\value{
Data frame with correlation results
}
\description{
Calculate correlations for genes as a module/anchor
}
