% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cache_init.R
\name{compute_base_hash}
\alias{compute_base_hash}
\title{Compute hash of base DAF for cache invalidation}
\usage{
compute_base_hash(base_daf)
}
\arguments{
\item{base_daf}{Base DAF object}
}
\value{
Hash string based on axis lengths and version
}
\description{
Compute hash of base DAF for cache invalidation
}
