% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts.R
\name{contract_to_markdown}
\alias{contract_to_markdown}
\title{Generate contract documentation as markdown}
\usage{
contract_to_markdown(contract)
}
\arguments{
\item{contract}{Contract to document}
}
\value{
Character string with markdown content
}
\description{
Generate contract documentation as markdown
}
