% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_data.R
\name{convert_daf_flagged_genes}
\alias{convert_daf_flagged_genes}
\title{Convert DAF flagged genes to character vector}
\usage{
convert_daf_flagged_genes(daf_obj, flag_name)
}
\arguments{
\item{daf_obj}{DAF object}

\item{flag_name}{Name of the boolean flag vector}
}
\value{
Character vector of gene names where flag is TRUE
}
\description{
Generic function to get genes where a boolean flag is TRUE.
Replaces convert_daf_lateral_genes, convert_daf_noisy_genes.
}
