% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_data.R
\name{convert_daf_to_mcview}
\alias{convert_daf_to_mcview}
\title{Convert DAF data to MCView format}
\usage{
convert_daf_to_mcview(daf_obj, var_name, atlas = FALSE)
}
\arguments{
\item{daf_obj}{DAF object}

\item{var_name}{Variable name to convert}

\item{atlas}{Whether to use atlas data}
}
\value{
Converted data in MCView format
}
\description{
Convert DAF data to MCView format
}
