% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cache_init.R
\name{copy_axes_to_cache}
\alias{copy_axes_to_cache}
\title{Copy axes from base DAF to cache DAF}
\usage{
copy_axes_to_cache(base_daf, cache_daf)
}
\arguments{
\item{base_daf}{Source DAF}

\item{cache_daf}{Target cache DAF}
}
\description{
Copy axes from base DAF to cache DAF
}
