% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_bundle}
\alias{create_bundle}
\title{Generate a 'deployment ready' bundle of the a project app}
\usage{
create_bundle(project, path = getwd(), name = "MCView_bundle")
}
\arguments{
\item{project}{path to the project directory}

\item{path}{path in which to create the bundle.}

\item{name}{name of the folder in which to create the bundle. The bundle would be created at \code{path}/\code{name}}
}
\description{
This would create a minimal shiny app in \code{path}/\code{name} directory which would contain:
\itemize{
\item{}{app.R file. }
\item{}{project config and cache. }
}
The bundle can then be deployed in shiny-server, shinyapps.io or any other environment that supports serving shiny apps.
Note: when deploying to these services - make sure you have the MCView package installed.
}
\examples{
\dontrun{
MCView::create_bundle(project = "PBMC", path = getwd(), name = "PBMC")
}

}
