% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cache_init.R
\name{create_files_cache_daf}
\alias{create_files_cache_daf}
\title{Create files-based cache DAF}
\usage{
create_files_cache_daf(base_daf, cache_path, base_path)
}
\arguments{
\item{base_daf}{Base DAF to reference}

\item{cache_path}{Path to cache directory}

\item{base_path}{Path to base DAF (for relative path)}
}
\value{
Files DAF object with base_daf_repository set
}
\description{
Create files-based cache DAF
}
