% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_core.R
\name{create_mcview_config_template}
\alias{create_mcview_config_template}
\title{Create template YAML config file}
\usage{
create_mcview_config_template(output_file)
}
\arguments{
\item{output_file}{Path to output YAML file}
}
\description{
Create template YAML config file
}
