% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cache.R
\name{create_memory_cache_daf}
\alias{create_memory_cache_daf}
\title{Create in-memory cache DAF}
\usage{
create_memory_cache_daf(base_daf, name = "cache")
}
\arguments{
\item{base_daf}{Base DAF to copy axes from}

\item{name}{Name for the memory DAF}
}
\value{
Memory DAF object with axes copied from base
}
\description{
Create in-memory cache DAF
}
