% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{Create a configuration folder for a project}
\usage{
create_project(project, project_dir = NULL)
}
\arguments{
\item{project}{name of the project}

\item{project_dir}{directory of the project. Would be created if not existing. If \code{NULL} - the project would be created under current directory.}
}
\description{
This would create a project directory with the default configuration files and directory structure.
A text editor would be opened in order to edit the project \code{config.yaml} file.
}
