% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_accessors.R
\name{daf_mat}
\alias{daf_mat}
\title{Safe DAF matrix accessor}
\usage{
daf_mat(daf_obj, axis1, axis2, name, required = TRUE)
}
\arguments{
\item{daf_obj}{DAF object}

\item{axis1}{First axis name}

\item{axis2}{Second axis name}

\item{name}{Matrix name}

\item{required}{If TRUE, throws error when missing}
}
\value{
Matrix or NULL
}
\description{
Safe DAF matrix accessor
}
