% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_queries.R
\name{daf_query_2d_coords}
\alias{daf_query_2d_coords}
\title{Retrieve 2D coordinates via DAF}
\usage{
daf_query_2d_coords(daf_obj, metacells = NULL)
}
\arguments{
\item{daf_obj}{DAF object}

\item{metacells}{Optional vector of metacell names to retrieve}
}
\value{
Data frame with metacell, x, y columns
}
\description{
Retrieve 2D coordinates via DAF
}
