% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_accessors.R
\name{daf_query_flagged_genes}
\alias{daf_query_flagged_genes}
\title{Query genes by a boolean flag vector}
\usage{
daf_query_flagged_genes(daf_obj, flag_name)
}
\arguments{
\item{daf_obj}{DAF object}

\item{flag_name}{Name of the boolean flag vector (e.g., "is_marker", "is_lateral")}
}
\value{
Character vector of gene names where flag is TRUE
}
\description{
Generic function to query genes where a boolean flag is TRUE.
Replaces daf_query_marker_genes, daf_query_lateral_genes, daf_query_noisy_genes.
}
