% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_data.R
\name{daf_query_gene_max_umis}
\alias{daf_query_gene_max_umis}
\title{Get max UMIs per gene using DAF query}
\usage{
daf_query_gene_max_umis(daf_obj)
}
\arguments{
\item{daf_obj}{DAF object}
}
\value{
Named vector of max UMIs per gene
}
\description{
Get max UMIs per gene using DAF query
}
