% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_queries.R
\name{daf_query_gene_sum_umis}
\alias{daf_query_gene_sum_umis}
\title{Get sum UMIs per gene using DAF query}
\usage{
daf_query_gene_sum_umis(daf_obj)
}
\arguments{
\item{daf_obj}{DAF object}
}
\value{
Named vector of sum UMIs per gene
}
\description{
Get sum UMIs per gene using DAF query
}
