% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_queries.R
\name{daf_query_lateral_genes}
\alias{daf_query_lateral_genes}
\title{Get lateral genes using DAF query}
\usage{
daf_query_lateral_genes(daf_obj)
}
\arguments{
\item{daf_obj}{DAF object}
}
\value{
Character vector of lateral gene names
}
\description{
Returns genes where is_lateral is TRUE
}
