% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_data.R
\name{daf_query_marker_genes}
\alias{daf_query_marker_genes}
\title{Get marker genes using DAF query}
\usage{
daf_query_marker_genes(daf_obj)
}
\arguments{
\item{daf_obj}{DAF object}
}
\value{
Character vector of marker gene names
}
\description{
Returns genes where is_marker is TRUE
}
