% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_queries.R
\name{daf_query_mc_mat}
\alias{daf_query_mc_mat}
\title{Retrieve a slice of the UMIs matrix via DAF}
\usage{
daf_query_mc_mat(daf_obj, genes = NULL, metacells = NULL)
}
\arguments{
\item{daf_obj}{DAF object}

\item{genes}{Optional vector of gene names to retrieve}

\item{metacells}{Optional vector of metacell names to retrieve}
}
\value{
Matrix slice with genes as rows and metacells as columns
}
\description{
Retrieves required genes/metacells from the UMIs matrix.
Filtering is done in R after retrieval.
}
