% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_queries.R
\name{daf_query_mc_sum}
\alias{daf_query_mc_sum}
\title{Retrieve total UMIs per metacell via DAF}
\usage{
daf_query_mc_sum(daf_obj, metacells = NULL)
}
\arguments{
\item{daf_obj}{DAF object}

\item{metacells}{Optional vector of metacell names to retrieve}
}
\value{
Named vector of total UMIs per metacell
}
\description{
Retrieve total UMIs per metacell via DAF
}
