% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_data.R
\name{daf_query_type_counts}
\alias{daf_query_type_counts}
\title{Get metacell count per type using DAF query}
\usage{
daf_query_type_counts(daf_obj)
}
\arguments{
\item{daf_obj}{DAF object}
}
\value{
Named vector of counts per type
}
\description{
Get metacell count per type using DAF query
}
