% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_data.R
\name{daf_scalar}
\alias{daf_scalar}
\title{Safe DAF scalar accessor}
\usage{
daf_scalar(daf_obj, name, default = NULL)
}
\arguments{
\item{daf_obj}{DAF object}

\item{name}{Scalar name}

\item{default}{Default value if missing}
}
\value{
Scalar value or default
}
\description{
Safe DAF scalar accessor
}
