% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_accessors.R
\name{daf_vec}
\alias{daf_vec}
\title{Safe DAF vector accessor}
\usage{
daf_vec(daf_obj, axis, name, required = TRUE)
}
\arguments{
\item{daf_obj}{DAF object}

\item{axis}{Axis name}

\item{name}{Property name}

\item{required}{If TRUE, throws error when missing}
}
\value{
Vector or NULL
}
\description{
Safe DAF vector accessor
}
