% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{dataset_ls}
\alias{dataset_ls}
\title{List available datasets}
\usage{
dataset_ls()
}
\value{
Character vector of dataset names
}
\description{
List available datasets
}
\examples{
\dontrun{
dataset_ls()
}

}
