% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{dataset_rm}
\alias{dataset_rm}
\title{Remove a dataset from a project}
\usage{
dataset_rm(dataset)
}
\arguments{
\item{dataset}{name of the dataset to remove}

\item{project}{path to the project directory}
}
\description{
Remove a dataset from a project
}
\examples{
\dontrun{
dataset_rm("PBMC", "PBMC163k")
}

}
