% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_core.R
\name{extract_config_for_daf}
\alias{extract_config_for_daf}
\title{Extract configuration for DAF mode}
\usage{
extract_config_for_daf(daf_obj, dataset_name, config_file = NULL)
}
\arguments{
\item{daf_obj}{DAF object}

\item{dataset_name}{Default dataset name}

\item{config_file}{Optional YAML config file}
}
\description{
Extract configuration for DAF mode
}
