% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{gene_names}
\alias{gene_names}
\title{Get gene names from DAF}
\usage{
gene_names(dataset, atlas = FALSE)
}
\arguments{
\item{dataset}{Dataset name}

\item{atlas}{Whether to use atlas data}
}
\value{
Character vector of gene names
}
\description{
Get gene names from DAF
}
