% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_all_tab_names}
\alias{get_all_tab_names}
\title{Get all possible tab names from tab definitions}
\usage{
get_all_tab_names()
}
\value{
Character vector of all tab names
}
\description{
Get all possible tab names from tab definitions
}
