% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cache.R
\name{get_base_daf}
\alias{get_base_daf}
\title{Get base DAF for a dataset (without cache)}
\usage{
get_base_daf(dataset)
}
\arguments{
\item{dataset}{Dataset name}
}
\value{
Base DAF object or NULL
}
\description{
Get base DAF for a dataset (without cache)
}
