% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cache_validity.R
\name{get_cache_daf}
\alias{get_cache_daf}
\title{Get cache DAF for a dataset}
\usage{
get_cache_daf(dataset)
}
\arguments{
\item{dataset}{Dataset name}
}
\value{
Cache DAF object (writable) or NULL
}
\description{
Get cache DAF for a dataset
}
