% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gene_expression.R
\name{get_cell_types_mat}
\alias{get_cell_types_mat}
\title{Get UMI matrix aggregated by cell types}
\usage{
get_cell_types_mat(
  cell_types,
  metacell_types,
  dataset,
  projected = FALSE,
  atlas = FALSE,
  corrected = FALSE
)
}
\arguments{
\item{cell_types}{Vector of cell type names}

\item{metacell_types}{Metacell types data frame}

\item{dataset}{Dataset name}

\item{projected}{Use projected data}

\item{atlas}{Whether to use atlas data}

\item{corrected}{Use corrected data}
}
\value{
UMI matrix (genes x cell_types)
}
\description{
Get UMI matrix aggregated by cell types
}
