% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cache_validity.R
\name{get_complete_daf}
\alias{get_complete_daf}
\title{Get complete DAF for a dataset (cache + base chained)}
\usage{
get_complete_daf(dataset)
}
\arguments{
\item{dataset}{Dataset name}
}
\value{
Complete DAF object for reading (cache has priority)
}
\description{
Get complete DAF for a dataset (cache + base chained)
}
