% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_data.R
\name{get_daf_for_query}
\alias{get_daf_for_query}
\title{Get DAF object for query (dataset or atlas)}
\usage{
get_daf_for_query(dataset, atlas = FALSE)
}
\arguments{
\item{dataset}{Dataset name}

\item{atlas}{Whether to use atlas data (default: FALSE)}
}
\value{
DAF object or NULL if not available
}
\description{
Consolidates the repeated pattern of selecting between dataset and atlas DAF.
}
