% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gene_expression.R
\name{get_gene_egc}
\alias{get_gene_egc}
\title{Get EGC for a single gene}
\usage{
get_gene_egc(
  gene,
  dataset,
  projected = FALSE,
  atlas = FALSE,
  corrected = FALSE
)
}
\arguments{
\item{gene}{Gene name}

\item{dataset}{Dataset name}

\item{projected}{Use projected data (for atlas)}

\item{atlas}{Whether to use atlas data}

\item{corrected}{Use corrected data}
}
\value{
Named vector of EGC values per metacell
}
\description{
Get EGC for a single gene
}
