% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cache.R
\name{get_mc_data}
\alias{get_mc_data}
\title{Get metacell data from DAF}
\usage{
get_mc_data(dataset, var_name, atlas = FALSE)
}
\arguments{
\item{dataset}{Dataset name}

\item{var_name}{Variable name (e.g., "mc_mat", "mc_sum", "mc2d")}

\item{atlas}{Whether to get atlas data}
}
\value{
Data in MCView format, or NULL if not available
}
\description{
Get metacell data from DAF
}
