% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gene_expression.R
\name{get_mc_egc}
\alias{get_mc_egc}
\title{Get expression per gene per cell (EGC) matrix}
\usage{
get_mc_egc(dataset, genes = NULL, atlas = FALSE, metacells = NULL)
}
\arguments{
\item{dataset}{Dataset name}

\item{genes}{Optional vector of gene names to filter}

\item{atlas}{Whether to use atlas data}

\item{metacells}{Optional vector of metacell names to filter}
}
\value{
EGC matrix (genes x metacells)
}
\description{
Get expression per gene per cell (EGC) matrix
}
