% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gene_expression.R
\name{get_mc_gene_modules_egc}
\alias{get_mc_gene_modules_egc}
\title{Get gene modules EGC matrix}
\usage{
get_mc_gene_modules_egc(
  dataset,
  modules = NULL,
  gene_modules = NULL,
  atlas = FALSE
)
}
\arguments{
\item{dataset}{Dataset name}

\item{modules}{Optional vector of module names to filter}

\item{gene_modules}{Optional gene modules data frame}

\item{atlas}{Whether to use atlas data}
}
\value{
EGC matrix aggregated by gene modules
}
\description{
Get gene modules EGC matrix
}
