% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gene_expression.R
\name{get_metacells_egc}
\alias{get_metacells_egc}
\title{Get EGC for specific metacells}
\usage{
get_metacells_egc(
  metacells,
  dataset,
  projected = FALSE,
  atlas = FALSE,
  corrected = FALSE
)
}
\arguments{
\item{metacells}{Vector of metacell names}

\item{dataset}{Dataset name}

\item{projected}{Use projected data (for atlas)}

\item{atlas}{Whether to use atlas data}

\item{corrected}{Use corrected data}
}
\value{
EGC matrix (genes x metacells)
}
\description{
Get EGC for specific metacells
}
