% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gene_expression.R
\name{get_samples_mat}
\alias{get_samples_mat}
\title{Get samples matrix}
\usage{
get_samples_mat(cell_types, metacell_types, dataset)
}
\arguments{
\item{cell_types}{Vector of cell type names}

\item{metacell_types}{Metacell types data frame}

\item{dataset}{Dataset name}
}
\value{
Samples UMI matrix
}
\description{
Get samples matrix
}
