% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cells.R
\name{has_cell_gene_umis}
\alias{has_cell_gene_umis}
\title{Check if cell-level gene UMIs are available}
\usage{
has_cell_gene_umis(dataset)
}
\arguments{
\item{dataset}{Dataset name}
}
\value{
TRUE if a cells DAF with cell x gene UMIs matrix is available
}
\description{
Check if cell-level gene UMIs are available
}
