% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import a project to MCView}
\usage{
import(project)
}
\arguments{
\item{project}{path to the project}
}
\description{
This would read the project \code{config.yaml} file and would import the metacell datasets listed
within it to MCView. The result would be a directory for each dataset under \code{project/cache} which
would contain objects used by MCView shiny app.
}
\examples{
\dontrun{
dir.create("raw")
download.file("http://www.wisdom.weizmann.ac.il/~atanay/metac_data/PBMC_processed.tar.gz", "raw/#' PBMC_processed.tar.gz")
untar("raw/PBMC_processed.tar.gz", exdir = "raw")
create_project("PBMC")
import("PBMC")
}

}
