% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_dataset}
\alias{import_dataset}
\title{Import a dataset to an MCView project}
\usage{
import_dataset(
  project,
  dataset,
  anndata_file,
  cell_type_field = "cluster",
  metacell_types_file = NULL,
  cell_type_colors_file = NULL,
  metadata_fields = NULL,
  metadata = NULL,
  metadata_types = NULL,
  metadata_colors = NULL
)
}
\arguments{
\item{project}{path to the project}

\item{dataset}{name for the dataset, e.g. "PBMC"}

\item{anndata_file}{path to \code{h5ad} file which contains the output of metacell2 pipeline (metacells python package).}

\item{cell_type_field}{name of a field in the anndata \code{object$obs} which contains a cell type (optional).
If the field doesn't exist and \code{metacell_types_file} are missing, MCView would cluster the
metacell matrix using kmeans++ algorithm (from the \code{tglkmeans} package).
If \code{metacell_types} parameter is set this field is ignored.}

\item{metacell_types_file}{path to a tabular file (csv,tsv) with cell type assignement for
each metacell. The file should have a column named "metacell" with the metacell ids and another
column named "cell_type" or "cluster" with the cell type assignment. Metacell ids that do
not exists in the data would be ignored.
If this parameter and \code{cell_type_field} are missing, MCView would cluster the
metacell matrix using kmeans++ algorithm (from the \code{tglkmeans} package).}

\item{cell_type_colors_file}{path to a tabular file (csv,tsv) with color assignement for
each cell type. The file should have a column named "cell_type" or "cluster" with the
cell types and another column named "color" with the color assignment. Cell types that do not
exist in the metacell types would be ignored.
If this is missing, MCView would use the \code{chameleon} package to assign a color for each cell type.}

\item{metadata_fields}{names of fields in the anndata \code{object$obs} which contains metadata for each metacell. 
By default character fields are treated as categorical (factors) and numeric fields as continuous. 
You can change the default behavior by setting \code{metadata_types} explicitly.}

\item{metadata}{can be either a data frame with a column named "metacell" with the metacell id and other metadata columns, 
or a name of a delimited file which contains such data frame. 
By default character fields are treated as categorical (factors) and numeric fields as continuous.
You can change the default behavior by setting \code{metadata_types} explicitly.}

\item{metadata_types}{a named vector with type per metadata field. Types can be "categorical" or "continuous".}

\item{metadata_colors}{a named list with colors for each metadata column. For categorical data the list element should 
contain a named vector with color per category, and for continuous field it should contain a list where the first element 
is a vector of breaks and the second element is a vector of colors.}
}
\description{
Read an \code{anndata} file which is the output of python \code{metacells} package,
and import the metacell dataset to MCView. Each project can have multiple datasets
which can be in the app using the right sidebar.
}
\details{
The function would create a directory under \code{project/cache/dataset} which
would contain objects used by MCView shiny app (such as the metacell matrix).

In addition, you can supply file with type assignment for each metacell
(\code{metacell_types_file}) and a file with color assignment for each metacell type
(\code{cell_type_colors_file}).
}
\examples{
\dontrun{
dir.create("raw")
download.file(
    "http://www.wisdom.weizmann.ac.il/~atanay/metac_data/PBMC_processed.tar.gz",
    "raw/PBMC_processed.tar.gz"
)
untar("raw/PBMC_processed.tar.gz", exdir = "raw")
create_project("PBMC")
import_dataset("PBMC", "PBMC163k", "raw/metacells.h5ad")
}

}
