% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metacell1.R
\name{import_dataset_metacell1}
\alias{import_dataset_metacell1}
\title{Import a dataset to an MCView project from metacell R package}
\usage{
import_dataset_metacell1(
  project,
  dataset,
  scdb,
  matrix,
  mc,
  mc2d,
  metacell_types_file,
  cell_type_colors_file,
  network = NULL,
  time_annotation_file = NULL,
  time_bin_field = NULL
)
}
\arguments{
\item{project}{path to the project}

\item{dataset}{name for the dataset, e.g. "PBMC"}

\item{scdb}{path to R metacell single cell RNA database}

\item{matrix}{name of the umi matrix to use}

\item{mc}{name of the metacell object to use}

\item{mc2d}{name of the 2d projection object to use}

\item{metacell_types_file}{path to a tabular file (csv,tsv) with cell type assignement for
each metacell. The file should have a column named "metacell" with the metacell ids and another
column named "cell_type" or "cluster" with the cell type assignment. Metacell ids that do
not exists in the data would be ignored. In addition, the file can have a column named
"age" or "mc_age" with age metadata per metacell}

\item{cell_type_colors_file}{path to a tabular file (csv,tsv) with color assignement for
each cell type. The file should have a column named "cell_type" or "cluster" with the
cell types and another column named "color" with the color assignment. Cell types that do not
exist in the metacell types would be ignored.}

\item{network}{name of the network object to use (optional)}

\item{time_annotation_file}{file with names for time bins (optional, only relevant with networks/flows). Should have a field named "time_bin" with the time bin id and another field named "time_desc" which contains the description of the time bin}

\item{time_bin_field}{name of a field in \code{cell_metadata} which contains time bin per cell (optional)}
}
\description{
Read objects from \code{metacell} R package and import a
metacell dataset to MCView.
}
\details{
The result would be a directory under \code{project/cache/dataset} which
would contain objects used by MCView shiny app (such as the metacell matrix).

In addition, you can supply file with type assignment for each metacell
(\code{metacell_types_file}) and a file with color assignment for each metacell type
(\code{cell_type_colors_file}).

Make sure that you have the R \code{metacell} package installed in order to use
this function.

\code{network}, \code{time_annotation_file} and \code{time_bin_field} are only relevant
if you computed flows/networks for your dataset and therefore are optional.
}
\examples{
\dontrun{
import_dataset_metacell1(
    "embflow",
    "153embs",
    scdb = "raw/scrna_db",
    matrix = "embs",
    mc = "embs",
    mc2d = "embs",
    metacell_types_file = "raw/metacell-types.csv",
    cell_type_colors_file = "raw/cell-type-colors.csv",
    network = "embs",
    time_annotation_file = "raw/time-annot.tsv",
    time_bin_field = "age_group"
)
}

}
