% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_core.R
\name{init_atlas}
\alias{init_atlas}
\title{Initialize atlas for projection comparison}
\usage{
init_atlas(atlas_daf)
}
\arguments{
\item{atlas_daf}{Atlas DAF object}
}
\description{
Initialize atlas for projection comparison
}
