% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_helpers.R
\name{julia_helpers_ready}
\alias{julia_helpers_ready}
\title{Check if Julia helpers are ready}
\usage{
julia_helpers_ready()
}
\value{
Logical
}
\description{
Returns TRUE if Julia helpers are enabled and initialized.
}
