% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_core.R
\name{load_yaml_config_to_daf}
\alias{load_yaml_config_to_daf}
\title{Load YAML configuration into DAF object}
\usage{
load_yaml_config_to_daf(daf_obj, config_file)
}
\arguments{
\item{daf_obj}{DAF object}

\item{config_file}{Path to YAML config file}
}
\description{
Load YAML configuration into DAF object
}
