% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcview_env.R
\name{mcv_exists}
\alias{mcv_exists}
\title{Check if variable exists in MCView environment}
\usage{
mcv_exists(var_name)
}
\arguments{
\item{var_name}{Name of the variable}
}
\description{
Check if variable exists in MCView environment
}
