% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcview_env.R
\name{mcv_get}
\alias{mcv_get}
\title{Get value from MCView environment}
\usage{
mcv_get(var_name)
}
\arguments{
\item{var_name}{Name of the variable}
}
\value{
The value stored in the environment
}
\description{
Get value from MCView environment
}
