% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcview_env.R
\name{mcv_set}
\alias{mcv_set}
\title{Set value in MCView environment}
\usage{
mcv_set(var_name, value)
}
\arguments{
\item{var_name}{Name of the variable}

\item{value}{Value to set}
}
\description{
Set value in MCView environment
}
