% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts.R
\name{mcview_all_contracts}
\alias{mcview_all_contracts}
\title{Get all MCView contracts}
\usage{
mcview_all_contracts()
}
\value{
Named list of all contract definitions
}
\description{
Get all MCView contracts
}
