% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts.R
\name{mcview_atlas_contract}
\alias{mcview_atlas_contract}
\title{Contract for Atlas tab}
\usage{
mcview_atlas_contract()
}
\description{
The query DAF should contain projection metadata, while the atlas DAF
(loaded via init_atlas) contains the reference data.
}
