% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts.R
\name{mcview_diff_expr_contract}
\alias{mcview_diff_expr_contract}
\title{Contract for Differential Expression tab}
\usage{
mcview_diff_expr_contract()
}
\description{
Contract for Differential Expression tab
}
