% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts.R
\name{mcview_gene_correlations_contract}
\alias{mcview_gene_correlations_contract}
\title{Contract for gene correlations data}
\usage{
mcview_gene_correlations_contract()
}
\description{
Contract for gene correlations data
}
