% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts.R
\name{mcview_gene_zero_fold_contract}
\alias{mcview_gene_zero_fold_contract}
\title{Contract for gene zero-fold data}
\usage{
mcview_gene_zero_fold_contract()
}
\description{
Contract for gene zero-fold data
}
