% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts.R
\name{mcview_markers_contract}
\alias{mcview_markers_contract}
\title{Contract for Markers tab}
\usage{
mcview_markers_contract()
}
\description{
Marker gene heatmap.
- Requires: core + is_marker [input]
- Enriched by: is_lateral, is_noisy [input], marker_rank [input]
}
