% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts_definitions.R
\name{mcview_projection_qc_contract}
\alias{mcview_projection_qc_contract}
\title{Contract for Projection QC tab}
\usage{
mcview_projection_qc_contract()
}
\description{
Atlas projection quality control.
- Requires: core + projection vectors [input]
- Accepts both similar and is_similar
- Accepts both projected_fraction and projected_geomean_fraction
}
