% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts_definitions.R
\name{mcview_tab_contract}
\alias{mcview_tab_contract}
\title{Get contract for a specific tab}
\usage{
mcview_tab_contract(tab_name)
}
\arguments{
\item{tab_name}{Name of the tab}
}
\value{
Contract definition or NULL if not found
}
\description{
Get contract for a specific tab
}
