% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gene_correlation.R
\name{mod_gene_correlation_server}
\alias{mod_gene_correlation_server}
\title{gene_correlation Server Function}
\usage{
mod_gene_correlation_server(
  id,
  dataset,
  metacell_types,
  cell_type_colors,
  gene_modules,
  globals
)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}

\item{dataset}{Reactive returning the current dataset name.}

\item{metacell_types}{Reactive returning metacell-type mapping data frame.}

\item{cell_type_colors}{Reactive returning cell-type colour data frame.}

\item{gene_modules}{Reactive returning gene-module mapping data frame.}

\item{globals}{Reactive values shared across modules.}
}
\description{
gene_correlation Server Function
}
