% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cache.R
\name{populate_dataset_cache}
\alias{populate_dataset_cache}
\title{Populate cache for a dataset in a running MCView session}
\usage{
populate_dataset_cache(dataset, what = NULL, force = FALSE, verbose = TRUE)
}
\arguments{
\item{dataset}{Dataset name}

\item{what}{What to cache (NULL = all)}

\item{force}{Recompute even if present}

\item{verbose}{Print progress}
}
\description{
Populate cache for a dataset in a running MCView session
}
