% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_heatmap.R
\name{print.gt_custom}
\alias{print.gt_custom}
\title{Print method for "gt_custom" class
we are overriding the print function, similiar to the one at shiny/render-plot.R:
https://github.com/rstudio/shiny/blob/main/R/render-plot.R
in order to provide a separate ggplot_build and gtable objects}
\usage{
\method{print}{gt_custom}(x)
}
\arguments{
\item{x}{An object of class "gt_custom"}
}
\description{
Print method for "gt_custom" class
we are overriding the print function, similiar to the one at shiny/render-plot.R:
https://github.com/rstudio/shiny/blob/main/R/render-plot.R
in order to provide a separate ggplot_build and gtable objects
}
