% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts_docs.R
\name{print_contract}
\alias{print_contract}
\title{Print contract summary}
\usage{
print_contract(contract)
}
\arguments{
\item{contract}{Contract to summarize}
}
\description{
Prints a human-readable summary of a contract.
}
