% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cache_init.R
\name{resolve_cache_path}
\alias{resolve_cache_path}
\title{Resolve cache path from configuration}
\usage{
resolve_cache_path(cache_dir, base_path, dataset_name)
}
\arguments{
\item{cache_dir}{Cache directory from config (relative or absolute)}

\item{base_path}{Path to base DAF}

\item{dataset_name}{Dataset name}
}
\value{
Absolute path to cache directory
}
\description{
Resolve cache path from configuration
}
