% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcview_env.R
\name{set_backend}
\alias{set_backend}
\title{Set backend mode}
\usage{
set_backend(kind, path = NULL, daf_obj = NULL)
}
\arguments{
\item{kind}{Backend type: "project" or "daf"}

\item{path}{Project path (for project mode)}

\item{daf_obj}{DAF object (for DAF mode)}
}
\description{
Set backend mode
}
