% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cache_init.R
\name{set_cache_metadata}
\alias{set_cache_metadata}
\title{Set cache metadata scalars}
\usage{
set_cache_metadata(cache_daf, base_daf)
}
\arguments{
\item{cache_daf}{Cache DAF to update}

\item{base_daf}{Base DAF to read version from}
}
\description{
Set cache metadata scalars
}
