% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{set_default_project}
\alias{set_default_project}
\title{Set default project}
\usage{
set_default_project(project)
}
\arguments{
\item{project}{path to the project directory}
}
\description{
Create a symbolic link to the \code{project} in order to run MCView without
specifying a project directory.
}
\details{
Note that in order for this to work you must have
write permissions in the installation directory of MCView. Used mainly for debugging purposes.
}
