% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{update_2d_projection}
\alias{update_2d_projection}
\title{Update default 2D projection for a dataset}
\usage{
update_2d_projection(project, dataset, layout, graph)
}
\arguments{
\item{project}{path to the project}

\item{dataset}{name for the dataset, e.g. "PBMC". The name of the dataset can only contain alphanumeric characters, dots, dashes and underscores.}

\item{layout}{a data frame with a column named "metacell" with the metacell id and other columns with the x and y coordinates of the metacell.}

\item{graph}{a data frame with a column named "from", "to" and "weight" with the ids of the metacells and the weight of the edge. If NULL, the graph would be taken from the anndata object.}
}
\description{
Update default 2D projection for a dataset
}
