% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_cache.R
\name{update_cache_metadata}
\alias{update_cache_metadata}
\title{Update cache metadata after population}
\usage{
update_cache_metadata(cache_daf, base_daf)
}
\arguments{
\item{cache_daf}{Cache DAF}

\item{base_daf}{Base DAF}
}
\description{
Update cache metadata after population
}
