% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gene_modules.R
\name{update_gene_modules}
\alias{update_gene_modules}
\title{Update gene modules}
\usage{
update_gene_modules(project, dataset, gene_modules_file)
}
\arguments{
\item{project}{path to the project}

\item{dataset}{name for the dataset, e.g. "PBMC"}

\item{gene_modules_file}{path to a tabular file (csv,tsv) with assignment of genes to gene modules. Should have a field named "gene" with the gene name and a field named "module" with the name of the gene module.}
}
\description{
Change the gene modules of an MCView app.
}
\details{
This is usually done after a first iteration of annotation using the "Gene modules" tab in the MCView annotation, which can export a valid \code{gene_modules_file}.
The file should have a column named "gene" with the gene names and another column named "module" with the id of the gene module.
Note that the exported file from the __MCView__ app might contain additional fields which will be ignored in this function.

Under the hood - MCView updates a file named "gene_modules.tsv" under \code{project/cache/dataset}, which can also be edited manually.
}
\examples{
\dontrun{
update_gene_modules("PBMC", "PBMC163k", "raw/gene-modules.csv")
}

}
