% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_types.R
\name{update_metacell_types}
\alias{update_metacell_types}
\title{Update cell type assignment for each metacell}
\usage{
update_metacell_types(project, dataset, metacell_types_file)
}
\arguments{
\item{project}{path to the project directory}

\item{dataset}{name for the dataset, e.g. "PBMC"}

\item{metacell_types_file}{path to a tabular file (csv,tsv) with cell type assignement for
each metacell. The file should have a column named "metacell" with the metacell ids and another
column named "cell_type" or "cluster" with the cell type assignment. Metacell ids that do
not exists in the data would be ignored.}
}
\description{
Change the cell type assignments for each metacell to the ones listed at \code{metacell_types_file}.
}
\details{
This is usually done after a first iteration of annotation using the "Annotate" tab in the MCView annotation, which can export a valid \code{metacell_types_file}.
The file should have a column named "metacell" with the metacell ids and another
column named "cell_type" or "cluster" with the cell type assignment.

Note that the exported file from the __MCView__ app contains additional fields
which will be ignored in this function.

Under the hood - MCView updates a file named "metacell_types.tsv" under \code{project/cache/dataset}, which can also be edited manually.
}
\examples{
\dontrun{
update_metacell_types("PBMC", "PBMC163k", "raw/metacell-clusters.csv")
}

}
