% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gene_mc.R
\name{calc_gene_gene_correlations}
\alias{calc_gene_gene_correlations}
\title{Calculate gene-gene correlations (correlation between input genes)}
\usage{
calc_gene_gene_correlations(
  dataset,
  genes,
  cell_type_filter = NULL,
  threshold = 0,
  atlas = FALSE
)
}
\arguments{
\item{dataset}{Dataset name}

\item{genes}{Vector of gene names}

\item{cell_type_filter}{Optional vector of cell types to filter metacells}

\item{threshold}{Minimum correlation threshold to include}
}
\value{
Data frame with gene-gene correlation results
}
\description{
Calculate gene-gene correlations (correlation between input genes)
}
