% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{cell_metadata_to_metacell}
\alias{cell_metadata_to_metacell}
\title{Convert cell metadata to metacell metadata}
\usage{
cell_metadata_to_metacell()
}
\description{
Convert cell metadata to metacell metadata
}
