% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{Create a configuration folder for a project}
\usage{
create_project(project, project_dir = NULL, edit_config = TRUE)
}
\arguments{
\item{project}{path of the project}

\item{edit_config}{open file editor for config file editing}
}
\description{
Create a project directory with the default configuration files and directory structure.
A text editor would be opened in order to edit the project \code{config.yaml} file.
}
\examples{
\dontrun{
dir.create("raw")
download.file("http://www.wisdom.weizmann.ac.il/~atanay/metac_data/PBMC_processed.tar.gz", "raw/#' PBMC_processed.tar.gz")
untar("raw/PBMC_processed.tar.gz", exdir = "raw")
create_project("PBMC")
}

}
