% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{dataset_ls}
\alias{dataset_ls}
\title{List datasets in a project}
\usage{
dataset_ls(project)
}
\arguments{
\item{project}{path to the project directory}
}
\value{
names of the existing datasets in the project
}
\description{
List datasets in a project
}
\examples{
\dontrun{
dataset_ls("PBMC")
}

}
