% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{dataset_rm}
\alias{dataset_rm}
\title{Remove a dataset}
\usage{
dataset_rm(dataset)
}
\arguments{
\item{dataset}{name of the dataset to remove}
}
\description{
Remove a dataset
}
\examples{
\dontrun{
dataset_rm("PBMC163k")
}

}
