% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import a project to MCView}
\usage{
import(project)
}
\arguments{
\item{project}{path to the project}
}
\description{
This would read the project \code{config.yaml} file and would import the metacell datasets listed
within it to MCView. The result would be a directory for each dataset under \code{project/data} which
would contain objects used by MCView shiny app.
}
