% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_dataset}
\alias{import_dataset}
\title{Import a dataset to an MCView project}
\usage{
import_dataset(
  project,
  dataset,
  anndata_file,
  cell_type_field = "cluster",
  metacell_types_file = NULL,
  cell_type_colors_file = NULL,
  calc_gg_cor = TRUE
)
}
\arguments{
\item{project}{path to the project}

\item{dataset}{name for the dataset, e.g. "PBMC"}

\item{anndata_file}{path to \code{h5ad} file which contains the output of metacell2 pipeline (metacells python package).}

\item{cell_type_field}{name of a field in the anndata object$obs which contains a cell type (optional).
If the field doesn't exist and \code{metacell_types_file} are missing, MCView would cluster the
metacell matrix using kmeans++ algorithm (from the \code{tglkmeans} package).
If \code{metacell_types} parameter is set this field is ignored.}

\item{metacell_types_file}{path to a tabular file (csv,tsv) with cell type assignement for
each metacell. The file should have a column named "metacell" with the metacell ids and another
column named "cell_type" or "cluster" with the cell type assignment. Metacell ids that do
not exists in the data would be ignored.
If this parameter and \code{cell_type_field} are missing, MCView would cluster the
metacell matrix using kmeans++ algorithm (from the \code{tglkmeans} package).}

\item{cell_type_colors_file}{path to a tabular file (csv,tsv) with color assignement for
each cell type. The file should have a column named "cell_type" or "cluster" with the
cell types and another column named "color" with the color assignment. Cell types that do not
exist in the metacell types would be ignored.
If this is missing, MCView would use the \code{chameleon} package to assign a color for each cell type.}

\item{calc_gg_cor}{calculate top 30 correlated and anti-correlated genes for each gene. This computation can be heavy
for large datasets or weaker machines, so you can set \code{calc_gg_cor=FALSE} to skip it. Note that then this feature
would be missing from the app.}
}
\description{
Read an \code{anndata} file which is the output of python \code{metacells} package,
and import the metacell dataset to MCView. Each project can have multiple datasets
which can be in the app using the right sidebar.
}
\details{
The function would create a directory under \code{project/cache/dataset} which
would contain objects used by MCView shiny app (such as the metacell matrix).

In addition, you can supply file with type assignment for each metacell
(\code{metacell_types_file}) and a file with color assignment for each metacell type
(\code{cell_type_colors_file}).
}
\examples{
\dontrun{
dir.create("raw")
download.file(
    "http://www.wisdom.weizmann.ac.il/~atanay/metac_data/PBMC_processed.tar.gz",
    "raw/PBMC_processed.tar.gz"
)
untar("raw/PBMC_processed.tar.gz", exdir = "raw")
create_project("PBMC")
import_dataset("PBMC", "PBMC163k", "raw/metacells.h5ad")
}

}
