% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCView-package.R
\docType{package}
\name{MCView-package}
\alias{MCView}
\alias{MCView-package}
\title{MCView: A Shiny App for Metacell Analysis}
\description{
A shiny app for metacell analysis
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

}
\keyword{internal}
