% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts_definitions.R
\name{mcview_atlas_contract}
\alias{mcview_atlas_contract}
\title{Contract for Atlas tab}
\usage{
mcview_atlas_contract()
}
\description{
Atlas comparison visualization.
- Requires: core + projection metadata [input]
- Accepts both similar and is_similar
- Accepts both projected_fraction and projected_geomean_fraction
}
