% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daf_contracts.R
\name{mcview_core_contract}
\alias{mcview_core_contract}
\title{Get the core MCView contract}
\usage{
mcview_core_contract()
}
\value{
List containing axes, vectors, matrices, and scalars specifications
}
\description{
This contract specifies the minimum required data for MCView to function.
}
