% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gene_mc.R
\name{plot_correlation_barplot}
\alias{plot_correlation_barplot}
\title{Create correlation barplot for single gene}
\usage{
plot_correlation_barplot(correlation_results, gene)
}
\arguments{
\item{correlation_results}{Data frame from correlation calculations}

\item{gene}{Gene name to plot}
}
\description{
Create correlation barplot for single gene
}
