% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gene_mc.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{Create correlation heatmap for multiple genes}
\usage{
plot_correlation_heatmap(
  correlation_results,
  input_genes,
  dataset,
  threshold = 0.3,
  cluster = TRUE,
  max_genes = 50,
  mask_low_correlations = FALSE,
  correlation_mode = "individual"
)
}
\arguments{
\item{correlation_results}{Data frame from calc_individual_correlations}

\item{input_genes}{Vector of input gene names}

\item{dataset}{Dataset name}

\item{threshold}{Correlation threshold for filtering}

\item{cluster}{Whether to cluster genes}

\item{max_genes}{Maximum number of genes to include in heatmap}
}
\description{
Create correlation heatmap for multiple genes
}
