% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the MCView Application}
\usage{
run_app(project, port = NULL, host = NULL, launch.browser = FALSE, ...)
}
\arguments{
\item{project}{path of the project to run.}

\item{port}{app port}

\item{host}{app host}

\item{launch.browser}{launch web browser after app start}

\item{...}{
  Arguments passed on to \code{\link[shiny:shinyApp]{shiny::shinyApp}}
  \describe{
    \item{\code{ui}}{The UI definition of the app (for example, a call to
\code{fluidPage()} with nested controls).

If bookmarking is enabled (see \code{enableBookmarking}), this must be
a single argument function that returns the UI definition.}
    \item{\code{server}}{A function with three parameters: \code{input}, \code{output}, and
\code{session}. The function is called once for each session ensuring that each
app is independent.}
    \item{\code{onStart}}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}
    \item{\code{options}}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}
    \item{\code{uiPattern}}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}
    \item{\code{enableBookmarking}}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}
  }}
}
\description{
Loads project data and runs the MCView application.
}
\examples{
\dontrun{
MCView::run_app("PBMC")
MCView::run_app(project = "PBMC", port = 5555, host = "127.0.0.1")
}

}
