% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{update_metadata}
\alias{update_metadata}
\title{Update metadata for a dataset}
\usage{
update_metadata(
  project,
  dataset,
  metadata = NULL,
  metadata_fields = NULL,
  anndata_file = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{project}{path to the project}

\item{dataset}{name for the dataset, e.g. "PBMC"}

\item{metadata}{can be either a data frame with a column named "metacell" with the metacell id and other metadata columns
or a name of a delimited file which contains such data frame.}

\item{metadata_fields}{names of fields in the anndata \code{object$obs} which contains metadata for each metacell.
The fields should can be either numeric or categorical. You can use
\code{cell_metadata_to_metacell} with \code{categorical=TRUE} to convert from categorical to a numeric score (e.g. using fraction of the category). You can use 'all' in order to import all the fields
 of the anndata object.}

\item{anndata_file}{path to \code{h5ad} file which contains the output of metacell2 pipeline (metacells python package).}

\item{overwrite}{overwrite all existing metadata. If \code{FALSE} - would override only existing metadata fields.}
}
\description{
Update metadata for a dataset
}
